Represent colors and other values as uint8 instead of int - a 1.5x speedup in frame rendering
Buffer is being sent in binary over the websocket to avoid
json.Marshal's base64 encoding of uint8 values - it thinks they're bytes
and there's no simple way to override it - which also improves the web
client's performance slightly

+++ effects/race_test_effect.go
@@ -12,6 +12,6 @@ type RaceTestEffect struct {}
 func (e *RaceTestEffect) Render(s *strip.LEDStrip) {
 	s.Clear()
 	for i := range s.Buffer {
-		s.Buffer[i] = []int{255, 0, 0}
+		s.Buffer[i] = []uint8{255, 0, 0}
 	}
 }
